# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  messages.BulkSendStartResponse:
    properties:
      message:
        example: Bulk send started in background
        type: string
      success:
        example: true
        type: boolean
      total:
        example: 123
        type: integer
    type: object
  messages.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      error:
        example: Validation error
        type: string
      message:
        example: Подробное описание ошибки
        type: string
    type: object
  messages.SendMediaMessageRequest:
    properties:
      async:
        example: true
        type: boolean
      file_data:
        example: base64_encoded_data
        type: string
      filename:
        example: image.png
        type: string
      message:
        example: Сообщение с медиа
        type: string
      message_type:
        example: image
        type: string
      mime_type:
        example: image/png
        type: string
      phone_number:
        example: "79991234567"
        type: string
    required:
    - file_data
    - filename
    - message_type
    - mime_type
    - phone_number
    type: object
  messages.SendMessageRequest:
    properties:
      async:
        example: true
        type: boolean
      message:
        example: Привет! Это тестовое сообщение
        type: string
      phone_number:
        example: "79991234567"
        type: string
    required:
    - message
    - phone_number
    type: object
  messages.SendMessageResponse:
    properties:
      id:
        example: message_id_123
        type: string
      message:
        example: Message sent successfully
        type: string
      status:
        example: sent
        type: string
      success:
        example: true
        type: boolean
    type: object
  types.HealthResponse:
    properties:
      status:
        example: ok
        type: string
      time:
        example: "2023-01-01T12:00:00Z"
        type: string
    type: object
  types.StatusResponse:
    properties:
      status:
        example: running
        type: string
      timestamp:
        example: "2023-01-01T12:00:00Z"
        type: string
      version:
        example: 1.0.0
        type: string
    type: object
  types.SuccessResponse:
    properties:
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  types.WhatGateSettings:
    properties:
      api_key:
        example: your_api_key
        type: string
      base_url:
        example: https://whatsgate.ru/api/v1
        type: string
      whatsapp_id:
        example: your_whatsapp_id
        type: string
    required:
    - api_key
    - whatsapp_id
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Сервис для отправки сообщений через WhatsApp с использованием API WhatGate
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: WhatsApp Service API
  version: "1.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: Проверяет работоспособность сервиса
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.HealthResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Проверка работоспособности сервиса
      tags:
      - health
  /messages/bulk-send:
    post:
      consumes:
      - multipart/form-data
      description: Отправляет сообщения на несколько номеров. Поддерживает отправку
        медиа и текста. messages_per_hour обязателен и > 0. Все рассылки асинхронные.
        Тип сообщения для media определяется автоматически по mime_type.
      parameters:
      - description: Текст сообщения
        in: formData
        name: message
        required: true
        type: string
      - description: Асинхронно (true/false)
        in: formData
        name: async
        type: boolean
      - description: Сколько сообщений в час (обязателен, > 0)
        in: formData
        name: messages_per_hour
        required: true
        type: integer
      - description: Файл с номерами (xlsx)
        in: formData
        name: numbers_file
        required: true
        type: file
      - description: Медиа-файл (опционально, тип определяется по mime_type)
        in: formData
        name: media_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Запуск рассылки
          schema:
            $ref: '#/definitions/messages.BulkSendStartResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Массовая рассылка сообщений
      tags:
      - messages
  /messages/send:
    post:
      consumes:
      - application/json
      description: Отправляет текстовое сообщение через WhatsApp
      parameters:
      - description: Параметры сообщения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/messages.SendMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/messages.SendMessageResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Отправить текстовое сообщение
      tags:
      - messages
  /messages/send-media:
    post:
      consumes:
      - application/json
      description: Отправляет сообщение с медиа-файлом через WhatsApp
      parameters:
      - description: Параметры медиа-сообщения
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/messages.SendMediaMessageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/messages.SendMessageResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Отправить медиа-сообщение
      tags:
      - messages
  /messages/test-send:
    post:
      consumes:
      - multipart/form-data
      description: Отправляет тестовое сообщение на один номер, принимает multipart/form-data
        (phone, message, media_file)
      parameters:
      - description: Номер телефона
        in: formData
        name: phone
        required: true
        type: string
      - description: Текст сообщения
        in: formData
        name: message
        required: true
        type: string
      - description: Медиа-файл (опционально)
        in: formData
        name: media_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Успешный ответ
          schema:
            $ref: '#/definitions/messages.SendMessageResponse'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Тестовая отправка сообщения (текст/медиа)
      tags:
      - messages
  /settings:
    get:
      consumes:
      - application/json
      description: Возвращает текущие настройки WhatGate API
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.WhatGateSettings'
      summary: Получить настройки WhatGate
      tags:
      - settings
    put:
      consumes:
      - application/json
      description: Обновляет настройки WhatGate API
      parameters:
      - description: Настройки WhatGate
        in: body
        name: settings
        required: true
        schema:
          $ref: '#/definitions/types.WhatGateSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.WhatGateSettings'
        "400":
          description: Ошибка валидации
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Обновить настройки WhatGate
      tags:
      - settings
  /settings/reset:
    delete:
      consumes:
      - application/json
      description: Сбрасывает настройки WhatGate API (удаляет сохраненные данные)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.SuccessResponse'
        "500":
          description: Внутренняя ошибка сервера
          schema:
            $ref: '#/definitions/messages.ErrorResponse'
      summary: Сбросить настройки WhatGate
      tags:
      - settings
  /status:
    get:
      consumes:
      - application/json
      description: Возвращает текущий статус и информацию о сервисе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/types.StatusResponse'
      summary: Статус сервиса
      tags:
      - status
securityDefinitions:
  BearerAuth:
    description: 'Введите токен в формате: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"